+++ client/rest-high-level/src/test/java/org/opensearch/client/core/tasks/GetTaskResponseTests.java
@@ -95,6 +95,12 @@ public class GetTaskResponseTests extends OpenSearchTestCase {
         boolean cancellable = randomBoolean();
         boolean cancelled = cancellable == true ? randomBoolean() : false;
         TaskId parentTaskId = randomBoolean() ? TaskId.EMPTY_TASK_ID : randomTaskId();
+        long cancellationStartTime = -1;
+        long runningTimeSinceCancellationNanos = -1;
+        if (cancelled) {
+            cancellationStartTime = randomLong();
+            runningTimeSinceCancellationNanos = randomLong();
+        }
         Map<String, String> headers = randomBoolean()
             ? Collections.emptyMap()
             : Collections.singletonMap(randomAlphaOfLength(5), randomAlphaOfLength(5));
@@ -110,7 +116,9 @@ public class GetTaskResponseTests extends OpenSearchTestCase {
             cancelled,
             parentTaskId,
             headers,
-            randomResourceStats()
+            randomResourceStats(),
+            cancellationStartTime,
+            runningTimeSinceCancellationNanos
         );
     }