+++ server/src/internalClusterTest/java/org/opensearch/action/admin/cluster/node/tasks/TasksIT.java
@@ -544,9 +544,13 @@ public class TasksIT extends OpenSearchIntegTestCase {
         assertEquals(1, cancelTasksResponse.getTasks().size());
 
         // Tasks are marked as cancelled at this point but not yet completed.
-        List<TaskInfo> taskInfoList = client().admin().cluster().prepareListTasks()
-            .setActions(TestTaskPlugin.TestTaskAction.NAME + "*").get().getTasks();
-        for (TaskInfo taskInfo: taskInfoList) {
+        List<TaskInfo> taskInfoList = client().admin()
+            .cluster()
+            .prepareListTasks()
+            .setActions(TestTaskPlugin.TestTaskAction.NAME + "*")
+            .get()
+            .getTasks();
+        for (TaskInfo taskInfo : taskInfoList) {
             assertTrue(taskInfo.isCancelled());
             assertNotEquals(-1, taskInfo.getCancellationStartTime());
             assertNotEquals(-1, taskInfo.getRunningTimeSinceCancellationNanos());